\name{foo}
\alias{foo}
\alias{foobar}
\title{ Foo }
\description{
foo is a simple function that shows 
\describe{
\item{character}{shows the character}
\item{anything else}{shows that too}
}
More description.
}
\usage{
foo(x, a)
}

\arguments{
  \item{x}{ the thing to be tested}
  \item{a}{ something printed }
}
\details{
  There aren't many details for foo.
}
\value{
  Here you describe the output of foo.
}
\author{ 
  Eli Holmes, NOAA, Seattle, USA. 
  }

\seealso{ \code{\link{graywhales}} }

\references{
 Here you can put references.
}

\examples{
 foo("a","b")
 foo(1,"a")
}
