\name{graywhales}
\alias{graywhales}
\docType{data}
\title{ Gray Whale Data }
\description{
  Some data for Foo.
}
\usage{
data(graywhales)
}
\format{
The data are supplied as a matrix with years in the first column and counts in the second column.
}
\source{
\itemize{
\item{graywhales }{ Gerber L. R., Master D. P. D. and Kareiva P. M. (1999) Gray whales and the value of monitoring data in implementing the U.S. Endangered Species Act.  Conservation Biology, 13, 1215-1219.}
}
}
\examples{
str(graywhales)
}
\keyword{datasets}
